requirements.txt
backtrader
python-dotenv
scikit-learn
matplotlib
pandas
python-binance
setuptools
wheel
tensorflow
tqdm